editoast: move each function into the appropriate file
