--- editoast/src/modelsv2/rolling_stock_model/power_restriction.rs
@@ -1,14 +1,0 @@
-use diesel::sql_types::Text;
-use serde::Deserialize;
-use serde::Serialize;
-use utoipa::ToSchema;
-
-editoast_common::schemas! {
-    PowerRestriction,
-}
-
-#[derive(QueryableByName, Debug, Clone, Serialize, Deserialize, ToSchema)]
-pub struct PowerRestriction {
-    #[diesel(sql_type = Text)]
-    pub power_restriction: String,
-}
